Remove timm dependency
